Final changes made

+++ controller/adminController.js
@@ -2,22 +2,38 @@ const User = require('../models/User'),
       Book = require('../models/Book'),
       Request = require('../models/Request'),
       Issue = require('../models/Issue'),
-      Activity = require('../models/Activity');
+      Activity = require('../models/Activity'),
+      { validTime } = require('../utilities/validTime');
 
 class AdminController {
     constructor() {}
 
+    /**
+     * Renders the admin profile.
+     *
+     * @param {number} req Request object.
+     * @param {number} res Response object.
+     * @return {void} 
+     */
     async adminProfile(req, res, next) {
         try {
             
+            // Get all the books and the issue requests made to the admin
             const books = await Book.paginate({}, { page: 1, limit: 10 });
             const requests = await Request.find({});
              
+            // Render the profile
             return res.render('admin/profile', {
                 userInfo: req.user,
                 books: books.docs,
                 requests,
-                page: req.query.page
+                page: req.query.page,
+                error: {
+                    message: req.query.errMsg || ''
+                },
+                success: {
+                    message: req.query.scsMsg || ''
+                }
             });
 
         } catch (error) {
@@ -26,17 +42,33 @@ class AdminController {
         }
     }
 
+
+    /**
+     * Approves each issue request.
+     *
+     * @param {number} req Request object.
+     * @param {number} res Response object.
+     * @return {void} 
+     */
     async approveIssue(req, res, next) {
         const { user_id, book_id } = req.query;
 
         try {
+            // Get the the info about the book to be issued and also the user requesting for it
             const user = await User.findById({ _id: user_id });
             const book = await Book.findById({ _id: book_id });  
-            
-            if (book.stock < 1) {
-                return res.redirect('/admin/profile?page=requests');
+
+            // Only issue requests between 9 and 5 PM
+            if (!validTime()) {
+                return res.redirect('/admin/profile?page=requests&errMsg=Books can only be issued between 9 AM and 5 PM');
             }
             
+            // Make sure the book is in stock
+            if (book.stock < 1) {
+                return res.redirect('/admin/profile?page=requests&errMsg=Book out of stock');
+            }
+            
+            // Remove the request and add an issue for the user 
             await Request.findOneAndDelete({ 'user_id.id': user._id, 'book_info.id': book._id });
             let bookIndex = -1;
             user.requestDetails.forEach((element, index) => {
@@ -45,6 +77,7 @@ class AdminController {
                 }
             });
             user.requestDetails.splice(bookIndex, 1);
+            // Decrement the stock once the book is issued 
             book.stock -= 1;
 
             const { _id, title, author, bookId, stock } = book; 
@@ -62,7 +95,8 @@ class AdminController {
                 }
             });
             user.issueDetails.push(book._id);
-
+            
+            // Make a log of both activities
             const activity = new Activity({
                 info: {
                     id: book._id,
@@ -80,6 +114,7 @@ class AdminController {
                 }
             });
 
+            // Make all updates in parallel
             await Promise.all([
                 user.save(),
                 book.save(),
@@ -87,7 +122,7 @@ class AdminController {
                 activity.save()
             ]); 
 
-            res.redirect('/admin/profile?page=requests');
+            res.redirect('/admin/profile?page=requests&scsMsg=Book issued successfully');
 
         } catch (error) {
             console.log(error);
@@ -95,12 +130,22 @@ class AdminController {
         }
     }
 
+
+    /**
+     * Disapproves/rejects each issue request.
+     *
+     * @param {number} req Request object.
+     * @param {number} res Response object.
+     * @return {void} 
+     */
     async rejectIssue(req, res, next) {
         const { user_id, book_id } = req.query;
         try {
+            // Get the the info about the book to be issued and also the user requesting for it
             const user = await User.findById({ _id: user_id });
             const book = await Book.findById({ _id: book_id });   
             
+            // Find and remove the request
             await Request.findOneAndDelete({ 'user_id.id': user._id, 'book_info.id': book_id });
             let bookIndex = -1;
             user.requestDetails.forEach((element, index) => {
@@ -115,7 +160,7 @@ class AdminController {
                 book.save()
             ]); 
 
-            res.redirect('/admin/profile?page=requests');
+            res.redirect('/admin/profile?page=requests&scsMsg=Request rejected');
 
         } catch (error) {
             console.log(error);
